Fixed Makefile.in CPPFLAGS to include ${top_srcdir}, where CompilerSpecific.h is located.
No longer requires WarpUp includes to build (unless WarpUp support code
is to be compiled in, of course.)

Fixed a bug that made nasty clicks when a an end-of-sample had been reached.
It actually seems this bug was in version 4.180 as well.

+++ Device/requester.c
@@ -854,7 +854,7 @@ static void UpdateInfoWindow( struct AHIAudioModeRequesterExt *req )
         id);
     AddTail((struct List *) &req->InfoList,(struct Node *) &req->AttrNodes[i]);
     Sprintf(req->AttrNodes[i++].text, GetString(msgReqInfoResolution, req->Catalog),
-        bits, GetString((stereo ?
+        bits, (ULONG) GetString((stereo ?
           (pan ? msgReqInfoStereoPan : msgReqInfoStereo) :
           msgReqInfoMono), req->Catalog));
     AddTail((struct List *) &req->InfoList,(struct Node *) &req->AttrNodes[i]);
